+++ example/ck_tile/39_basic_gemm_softmax_topk/block_gemm_pipeline_agmem_bgmem_creg.hpp
@@ -6,9 +6,6 @@
 #include "block_gemm_pipeline_agmem_bgmem_creg_default_policy.hpp"
 
 #include "ck_tile/core.hpp"
-#include "ck_tile/core/tensor/tile_distribution.hpp"
-#include "ck_tile/ops/common.hpp"
-#include "ck_tile/ops/reduce.hpp"
 #include "ck_tile/ops/gemm/pipeline/gemm_pipeline_ag_bg_cr_scheduler.hpp"
 #include "ck_tile/ops/gemm/pipeline/gemm_pipeline_ag_bg_cr_base.hpp"
 
@@ -18,29 +15,20 @@ namespace ck_tile {
 //  B Tile Window: global memory
 //  C Distributed tensor: register
 template <typename Problem, typename Policy = ck_tile::BlockGemmPipelineAGmemBGmemCRegDefaultPolicy>
-struct BlockGemmSoftmaxTopkPipelineAGmemBGmemCReg
+struct BlockGemmPipelineAGmemBGmemCReg
 {
     using ADataType      = remove_cvref_t<typename Problem::ADataType>;
     using BDataType      = remove_cvref_t<typename Problem::BDataType>;
-    using WeightType     = remove_cvref_t<typename Problem::WeightType>;
-    using IndexType      = remove_cvref_t<typename Problem::IndexType>;
-    using ComputeDataType= float;
+    using CDataType      = remove_cvref_t<typename Problem::CDataType>;
     using BlockGemmShape = remove_cvref_t<typename Problem::BlockGemmShape>;
 
     static constexpr index_t kBlockSize = Problem::kBlockSize;
-    static constexpr index_t topk = Problem::topk;
 
     static constexpr index_t kMPerBlock = BlockGemmShape::kM;
     static constexpr index_t kNPerBlock = BlockGemmShape::kN;
     static constexpr index_t kKPerBlock = BlockGemmShape::kK;
 
     using BlockGemm = remove_cvref_t<decltype(Policy::template GetBlockGemm<Problem>())>;
-
-    struct ArgmaxPacket
-    {
-        WeightType arg;
-        IndexType value;
-    };
 
     CK_TILE_HOST_DEVICE static constexpr ck_tile::index_t GetStaticLdsSize()
     {
@@ -128,11 +116,11 @@ struct BlockGemmSoftmaxTopkPipelineAGmemBGmemCReg
 
         // stage 1
         // Separate this part?
-        // constexpr auto num_mfma_per_ds_read = sizeof(CDataType) / sizeof(ADataType) >
-        //                                               sizeof(CDataType) /
+        // constexpr auto num_mfma_per_ds_read = sizeof(ComputeDataType) / sizeof(ADataType) >
+        //                                               sizeof(ComputeDataType) /
         //                                               sizeof(BDataType)
-        //                                           ? sizeof(CDataType) /
-        //                                           sizeof(ADataType) : sizeof(CDataType)
+        //                                           ? sizeof(ComputeDataType) /
+        //                                           sizeof(ADataType) : sizeof(ComputeDataType)
         //                                           / sizeof(BDataType);
         constexpr auto num_mfma_stage1 = num_mfma_inst - (num_dsread_a_mfma + num_dsread_b_mfma);
         constexpr auto num_mfma_per_issue =
@@ -206,11 +194,9 @@ struct BlockGemmSoftmaxTopkPipelineAGmemBGmemCReg
     }
 #endif
 
-    template <typename ADramBlockWindowTmp, typename BDramBlockWindowTmp, typename ValueWindow, typename IndexWindow>
+    template <typename ADramBlockWindowTmp, typename BDramBlockWindowTmp>
     CK_TILE_HOST_DEVICE auto operator()(const ADramBlockWindowTmp& a_dram_block_window_tmp,
                                         const BDramBlockWindowTmp& b_dram_block_window_tmp,
-                                        const ValueWindow& value_dram_block_window,
-                                        const IndexWindow& index_dram_block_window,
                                         index_t num_loop,
                                         void* p_smem) const
     {
@@ -312,16 +298,6 @@ struct BlockGemmSoftmaxTopkPipelineAGmemBGmemCReg
 
         ABlockTile a_block_tile;
         BBlockTile b_block_tile;
-
-        using ValueBlockTileDistr = decltype(value_dram_block_window.get_tile_distribution());
-        using IndexBlockTileDistr = decltype(index_dram_block_window.get_tile_distribution());
-
-        using ValueBlockTile = decltype(make_static_distributed_tensor<WeightType>(ValueBlockTileDistr{}));
-        using IndexBlockTile = decltype(make_static_distributed_tensor<IndexType>(IndexBlockTileDistr{}));
-
-        ValueBlockTile value_block_tile;
-        IndexBlockTile index_block_tile;
-
         using ADramTileWindowStep = typename ADramBlockWindowTmp::BottomTensorIndex;
         using BDramTileWindowStep = typename BDramBlockWindowTmp::BottomTensorIndex;
         constexpr ADramTileWindowStep a_dram_tile_window_step = make_array(0, kKPerBlock);
@@ -429,120 +405,7 @@ struct BlockGemmSoftmaxTopkPipelineAGmemBGmemCReg
             iCounter--;
         }
 #endif
-        // apply softmax for c_block_tile
-        // reduction function for softmax
-        const auto f_max = [](auto e0, auto e1) { return max(e0, e1); };
-        const auto f_sum = [](auto e0, auto e1) { return e0 + e1; };
-
-        // m_local = rowmax(c_block_tile)
-        auto m_local = block_tile_reduce<ComputeDataType>(
-            c_block_tile, sequence<1>{}, f_max, std::numeric_limits<ComputeDataType>::lowest());
-        
-        block_tile_reduce_sync(m_local, f_max);
-
-        // Pcompute{j} = sum(exp(x - m_local))
-        auto p_compute =
-            make_static_distributed_tensor<ComputeDataType>(c_block_tile.get_tile_distribution());
-
-        constexpr auto p_spans = decltype(p_compute)::get_distributed_spans();
-
-        sweep_tile_span(p_spans[number<0>{}], [&](auto idx0) {
-            constexpr auto i_idx = make_tuple(idx0);
-
-            sweep_tile_span(p_spans[number<1>{}], [&](auto idx1) {
-                constexpr auto i_j_idx = make_tuple(idx0, idx1);
-
-                p_compute(i_j_idx) = exp(c_block_tile[i_j_idx] - m_local[i_idx]);
-            });
-        });
-
-        // rowsum for p_compute{i, j}
-        auto rowsum_p = block_tile_reduce<ComputeDataType>(
-            p_compute, sequence<1>{}, f_sum, ComputeDataType{0});
-
-        block_tile_reduce_sync(rowsum_p, f_sum);
-
-        // softmax = p_compute{i, j} / rowsum_p
-        sweep_tile_span(p_spans[number<0>{}], [&](auto idx0) {
-            constexpr auto i_idx = make_tuple(idx0);
-
-            sweep_tile_span(p_spans[number<1>{}], [&](auto idx1) {
-                constexpr auto i_j_idx = make_tuple(idx0, idx1);
-
-                p_compute(i_j_idx) = p_compute[i_j_idx] / rowsum_p[i_idx];
-            });
-        });
-
-        // apply topk for softmax output
-        auto x_tmp = p_compute;
-        // constexpr auto dst_dist = BlockGemmPipelineAGmemBGmemCRegDefaultPolicy::MakeOutputDistribution();
-
-        // argmax for topk
-        const auto f_argmax = [](ArgmaxPacket e0, ArgmaxPacket e1) {
-            return e0.arg > e1.arg ? e0 : e1;
-        };
-
-        for(index_t i_k = 0; i_k < topk; i_k++)
-        {
-            constexpr auto span_2d = decltype(p_compute)::get_distributed_spans();
-            auto packet            = [&]() {
-                auto tmp = make_static_distributed_tensor<ArgmaxPacket>(p_compute.get_tile_distribution());
-
-                sweep_tile_span(span_2d[number<0>{}], [&](auto idx0) {
-                    sweep_tile_span(span_2d[number<1>{}], [&](auto idx1) {
-                        const auto tile_idx = get_x_indices_from_distributed_indices(
-                            tmp.get_tile_distribution(), make_tuple(idx0, idx1));
-                        constexpr auto i_j_idx = make_tuple(idx0, idx1);
-                        ArgmaxPacket t;
-                        t.arg        = x_tmp(i_j_idx); // !!! we reference p_compute here
-                        t.value      = tile_idx.at(number<1>{});
-                        tmp(i_j_idx) = t;
-                    });
-                });
-                return tmp;
-            }();
-
-            auto argmax_init = ArgmaxPacket{-numeric<WeightType>::infinity(), 0};
-            auto r = block_tile_reduce<ArgmaxPacket>(packet, sequence<1>{}, f_argmax, argmax_init);
-            block_tile_reduce_xor_sync(r, f_argmax);
-
-            // auto value_block_tile = make_static_distributed_tensor<WeightType>(dst_dist);
-            // auto index_block_tile = make_static_distributed_tensor<IndexType>(dst_dist);
-
-            // Initialize value_block_tile and index_block_tile
-            tile_elementwise_inout([](auto& value) { value = 0; }, value_block_tile);
-            tile_elementwise_inout([](auto& index) { index = 0; }, index_block_tile);
-
-            sweep_tile_span(span_2d[number<0>{}], [&](auto idx0) {
-                sweep_tile_span(span_2d[number<1>{}], [&](auto idx1) {
-                    constexpr auto i_j_idx = make_tuple(idx0, idx1);
-                    ArgmaxPacket tmp       = r(i_j_idx);
-                    value_block_tile(i_j_idx)             = tmp.arg;
-                    index_block_tile(i_j_idx)             = tmp.value;
-                });
-            });
-
-            // update value
-            sweep_tile_span(span_2d[number<0>{}], [&](auto idx0) {
-                sweep_tile_span(span_2d[number<1>{}], [&](auto idx1) {
-                    const auto tile_idx = get_x_indices_from_distributed_indices(
-                        p_compute.get_tile_distribution(), make_tuple(idx0, idx1));
-                    auto col_id = tile_idx.at(number<1>{});
-
-                    constexpr auto i_j_idx = make_tuple(idx0, idx1);
-
-                    x_tmp(i_j_idx) = (col_id == r(i_j_idx).value) ? -numeric<WeightType>::infinity()
-                                                                    : x_tmp(i_j_idx);
-                });
-            });
-        }
-        // // cast to WeightType and apply CElementFunction
-        // value_block_tile = tile_elementwise_in(
-        //     [&](const auto& value) { return c_element_func(type_convert<WeightType>(value)); },
-        //     value_block_tile);
-
-        store_tile(value_dram_block_window, value_block_tile);
-        store_tile(index_dram_block_window, index_block_tile);
+        return c_block_tile;
     }
 };
